requirements.txt
amqp==5.2.0
asgiref==3.8.1
Babel==2.15.0
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
customtkinter==5.2.2
darkdetect==0.8.0
Django==5.1.8
django-filter==24.1
djangorestframework==3.15.2
django-cors-headers==4.7.0
drf-spectacular==0.27.1
gitdb==4.0.11
GitPython==3.1.43
gunicorn==23.0.0
idna==3.7
keyboard==0.13.5
kombu==5.4.0
lizard==1.17.10
maturin==1.7.4
packaging==24.1
pillow==10.4.0
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg2-binary==2.9.9
PyDriller==2.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
praw==7.7.1
pytz==2024.1
redis==5.1.1
regex==2024.5.15
requests==2.32.3
six==1.16.0
smmap==5.0.1
sqlparse==0.5.1
tkcalendar==1.6.1
tqdm==4.66.4
types-pytz==2024.1.0.20240417
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.0
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.3.1
